Rename landscape to synchronous and create statespace
The landscape module is being broken into several modules. First,
the StateSpace class is being moved to its own module, statespace.
The algorithms so far implemented in landscape are being moved to
the synchronous module. An asynchronous module is to come.
